+++ app/src/main/java/com/breadwallet/wallet/wallets/bitcoincash/WalletBchManager.java
@@ -62,6 +62,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.math.BigDecimal;
+import java.text.Normalizer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.Executor;
@@ -389,7 +390,7 @@ public class WalletBchManager extends BRCoreWalletManager implements BaseWalletM
     @Override
     public BigDecimal getFiatExchangeRate(Context app) {
         CurrencyEntity ent = CurrencyDataSource.getInstance(app).getCurrencyByCode(app, getIso(app), BRSharedPrefs.getPreferredFiatIso(app));
-        return new BigDecimal(ent == null ? 0 : (long) ent.rate); //dollars
+        return new BigDecimal(ent == null ? 0 : ent.rate); //dollars
     }
 
     @Override
